+++ src/JsonApi/Exception/ClientGeneratedIdNotSupported.php
@@ -6,7 +6,7 @@ namespace WoohooLabs\Yin\JsonApi\Exception;
 use WoohooLabs\Yin\JsonApi\Schema\Error;
 use WoohooLabs\Yin\JsonApi\Schema\ErrorSource;
 
-class ClientGeneratedIdNotSupported extends JsonApiException
+class ClientGeneratedIdNotSupported extends AbstractJsonApiException
 {
     /**
      * @var string