+++ src/JsonApi/Exception/JsonApiException.php
@@ -3,40 +3,9 @@ declare(strict_types=1);
 
 namespace WoohooLabs\Yin\JsonApi\Exception;
 
-use Exception;
-use WoohooLabs\Yin\JsonApi\Schema\Document\AbstractErrorDocument;
-use WoohooLabs\Yin\JsonApi\Schema\Document\ErrorDocument;
-use WoohooLabs\Yin\JsonApi\Schema\Error;
+class_alias(AbstractJsonApiException::class, 'WoohooLabs\Yin\JsonApi\Exception\JsonApiException');
 
-abstract class JsonApiException extends Exception implements JsonApiExceptionInterface
-{
-    /**
-     * @var AbstractErrorDocument
-     */
-    protected $errorDocument;
-
-    public function __construct(string $message = "", int $code = 0)
-    {
-        parent::__construct($message, $code);
-    }
-
-    protected function createErrorDocument(): AbstractErrorDocument
-    {
-        return new ErrorDocument();
-    }
-
-    /**
-     * @return Error[]
-     */
-    abstract protected function getErrors(): array;
-
-    public function getErrorDocument(): AbstractErrorDocument
-    {
-        $document = $this->createErrorDocument();
-        foreach ($this->getErrors() as $error) {
-            $document->addError($error);
-        }
-
-        return $document;
-    }
-}
+trigger_error(
+    'Class WoohooLabs\Yin\JsonApi\Exception\JsonApiException is deprecated, use ' . AbstractJsonApiException::class . ' instead.',
+    E_USER_DEPRECATED
+);