+++ src/JsonApi/Exception/QueryParamMalformed.php
@@ -6,7 +6,7 @@ namespace WoohooLabs\Yin\JsonApi\Exception;
 use WoohooLabs\Yin\JsonApi\Schema\Error;
 use WoohooLabs\Yin\JsonApi\Schema\ErrorSource;
 
-class QueryParamMalformed extends JsonApiException
+class QueryParamMalformed extends AbstractJsonApiException
 {
     /**
      * @var string