+++ src/JsonApi/Exception/ResourceIdMissing.php
@@ -6,7 +6,7 @@ namespace WoohooLabs\Yin\JsonApi\Exception;
 use WoohooLabs\Yin\JsonApi\Schema\Error;
 use WoohooLabs\Yin\JsonApi\Schema\ErrorSource;
 
-class ResourceIdMissing extends JsonApiException
+class ResourceIdMissing extends AbstractJsonApiException
 {
     public function __construct()
     {