+++ src/JsonApi/Exception/ResponseBodyInvalidJsonApi.php
@@ -9,7 +9,7 @@ use WoohooLabs\Yin\JsonApi\Schema\Document\ErrorDocument;
 use WoohooLabs\Yin\JsonApi\Schema\Error;
 use WoohooLabs\Yin\JsonApi\Schema\ErrorSource;
 
-class ResponseBodyInvalidJsonApi extends JsonApiException
+class ResponseBodyInvalidJsonApi extends AbstractJsonApiException
 {
     /**
      * @var ResponseInterface